feat: Add duration tracking to archive jobs
- Add database migration v22 to add duration_seconds REAL column to archive_jobs table
- Update ArchiveJob model to include duration_seconds field with subsecond precision
- Modify set_job_completed, set_job_failed, and set_job_skipped to calculate and store job duration
- Add Duration column to archive jobs table in web UI
- Implement format_duration helper to display durations in human-readable format:
  - Sub-second: "0.XYZs"
  - Under 1 minute: "X.Ys"
  - Under 1 hour: "Xm Y.Zs"
  - Over 1 hour: "Xh Ym Z.Zs"
- Update test fixtures to include new fields (duration_seconds, progress tracking fields)

Duration is automatically calculated using SQLite's julianday function when jobs complete.

+++ src/db/models.rs
@@ -502,8 +502,8 @@ pub struct ArchiveJob {
     pub error_message: Option<String>,
     pub metadata: Option<String>,
     pub created_at: String,
-    /// Duration of job execution in seconds (calculated from started_at to completed_at).
-    pub duration_seconds: Option<i64>,
+    /// Duration of job execution in seconds with subsecond accuracy (calculated from started_at to completed_at).
+    pub duration_seconds: Option<f64>,
 }
 
 impl ArchiveJob {
